requirements.txt
Django==5.2.5
djangorestframework==3.16.1
gunicorn==23.0.0
django-cors-headers==4.7.0
google-generativeai>=0.7.0
celery==5.3.6
redis==5.0.1   # broker
